Copy compose from deb96499dfe95073f5c1215c1287787683cb1e92

+++ compose/foundation/foundation/src/androidDeviceTest/kotlin/androidx/compose/foundation/pager/PagerNestedPrefetchingTest.kt
@@ -101,7 +101,7 @@ class PagerNestedPrefetchingTest(val config: ParamConfig) : BasePagerTest(config
             runBlocking {
                 // this will move the viewport so pages 1-2 are visible
                 // and schedule a prefetching for 3
-                listState.scrollBy(pageSizePx * 2f.toFloat())
+                listState.scrollBy(pageSizePx * 2f)
             }
         }
